+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+redis.om.spring.djl.enabled=true
+spring.main.allow-bean-definition-overriding=true
